projects/capstone/starter/requirements.txt
astroid==2.2.5
Click==8.1.7
ecdsa==0.13.2
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
future==0.17.1
isort==4.3.18
itsdangerous==2.1.2
Jinja2==3.0.1
lazy-object-proxy==1.4.0
MarkupSafe==2.1.3
mccabe==0.6.1
pycryptodome==3.3.1
pylint==2.3.1
python-jose-cryptodome==1.3.2
six==1.15.0
typed-ast==1.4.2
Werkzeug==2.0.0
wrapt==1.11.1
Flask-Cors==3.0.7
python-jose==3.3.0
SQLAlchemy==1.4.23